Make ``HashableDict`` impl more robust.
BEFORE: It attempted to get fancy by sorting the items of the dict before
converting to a list of tuples and hashing that. This could break if the
connection paramaters mix types like ``str`` and ``int`` (HINT: This happens
a fair bit in the wild).

AFTER: We just covert to tuples and hash; no more attempting to order.